small fix in integral

+++ rusted/integral/src/main.rs
@@ -10,41 +10,36 @@ fn f(x: f64, expr: &str) -> f64 {
     eval_str_with_context(expr, &context).unwrap()
 }
 
-fn lrect(a: f64, b: f64, n: u64, expr: &str) -> f64 {
+fn lrect(a: f64, b: f64, n: f64, expr: &str) -> f64 {
     let mut s: f64 = 0.0;
-    for i in 0..=n - 1 {
-        s += f(a + i as f64 * (b - a) / (n as f64), expr)
+    for i in 0..=(n - 1.0) as u64 {
+        s += f(a + i as f64 * (b - a) / n, expr)
     }
-    s * ((b - a) / n as f64)
+    s * ((b - a) / n)
 }
 
-fn rrect(a: f64, b: f64, n: u64, expr: &str) -> f64 {
+fn rrect(a: f64, b: f64, n: f64, expr: &str) -> f64 {
     let mut s: f64 = 0.0;
-    for i in 1..=n {
-        s += f(a + i as f64 * (b - a) / (n as f64), expr)
+    for i in 1..=n as u64 {
+        s += f(a + i as f64 * (b - a) / n, expr)
     }
-    s * ((b - a) / n as f64)
+    s * ((b - a) / n)
 }
 
-fn mrect(a: f64, b: f64, n: u64, expr: &str) -> f64 {
+fn mrect(a: f64, b: f64, n: f64, expr: &str) -> f64 {
     let mut s: f64 = 0.0;
-    for i in 0..=n - 1 {
-        s += f(
-            a + (b - a) * (2.0 * i as f64 + 1.0) / (2.0 * n as f64),
-            expr,
-        )
+    for i in 0..=(n - 1.0) as u64 {
+        s += f(a + (b - a) * (2.0 * i as f64 + 1.0) / (2.0 * n), expr)
     }
-    s * ((b - a) / n as f64)
+    s * ((b - a) / n)
 }
 
-fn trapezoid(a: f64, b: f64, n: u64, expr: &str) -> f64 {
+fn trapezoid(a: f64, b: f64, n: f64, expr: &str) -> f64 {
     let mut s: f64 = 0.0;
-    for i in 0..=n - 1 {
-        s += (f(a + i as f64 * (b - a) / n as f64, expr)
-            + f(a + (i + 1) as f64 * (b - a) / n as f64, expr))
-            / 2.0
+    for i in 0..=(n - 1.0) as u64 {
+        s += (f(a + i as f64 * (b - a) / n, expr) + f(a + (i + 1) as f64 * (b - a) / n, expr)) / 2.0
     }
-    s * ((b - a) / n as f64)
+    s * ((b - a) / n)
 }
 
 fn main() {
@@ -56,10 +51,10 @@ fn main() {
 
     let (a, b, eps) = scan_fmt!(&format!("{seg} {eps}"), "{} {} {}", f64, f64, f64).unwrap();
 
-    let (mut n, mut s1, mut s2): (u64, f64, f64) = (1, 0.0, f(a, &expr) * (b - a));
+    let (mut n, mut s1, mut s2): (f64, f64, f64) = (1.0, 0.0, f(a, &expr) * (b - a));
 
     while (s2 - s1).abs() > eps {
-        n *= 2;
+        n *= 2.0;
         (s1, s2) = (s2, lrect(a, b, n, &expr));
     }
 
@@ -68,7 +63,7 @@ fn main() {
     (s1, s2) = (0.0, f(a, &expr) * (b - a));
 
     while (s2 - s1).abs() > eps {
-        n *= 2;
+        n *= 2.0;
         (s1, s2) = (s2, rrect(a, b, n, &expr));
     }
 
@@ -77,7 +72,7 @@ fn main() {
     (s1, s2) = (0.0, f((a + b) / 2.0, &expr) * (b - a));
 
     while (s2 - s1).abs() > eps {
-        n *= 2;
+        n *= 2.0;
         (s1, s2) = (s2, mrect(a, b, n, &expr));
     }
 
@@ -86,7 +81,7 @@ fn main() {
     (s1, s2) = (0.0, (b - a) * (f(a, &expr) + f(b, &expr)) / 2.0);
 
     while (s2 - s1).abs() > eps {
-        n *= 2;
+        n *= 2.0;
         (s1, s2) = (s2, trapezoid(a, b, n, &expr));
     }
